REFACTOR: Created images folder
