+++ src/assets/images/home/pokedex-logo.png
