+++ src/assets/images/locations/waters-edge.jpg
