+++ src/components/pokemons/PokemonListCard.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import silouette from '@/assets/pokemons/silouette.png';
+import silouette from '@/assets/images/pokemons/silouette.png';
 import { pokemonColorTypes } from '@/constants/pokemonTypesColor';
 import store from '@/lib/store';
